requirements.txt
graphviz==0.20
importlib-metadata==4.12.0
Markdown==3.4.1
prettytable==3.3.0
pystache==0.6.0
PyYAML==6.0
wcwidth==0.2.5
zipp==3.8.1
